requirements.txt
Flask
flwr
numpy
scikit-learn
seaborn
matplotlib
Pillow